stereo-rdkx5/requirements.txt
torch
torchvision
torchaudio
numpy
scipy
opencv-python
matplotlib
tqdm
pyyaml
yacs
tensorboard
onnx
onnxruntime
onnxsim
rich
